Create presentation package and move ui components to this package
