+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.Calculator"
         tools:targetApi="33">
         <activity
-            android:name="com.sadegh.calculator.MainActivity"
+            android:name="com.sadegh.calculator.presentation.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.Calculator">
             <intent-filter>
