+++ app/src/main/java/com/sadegh/calculator/presentation/MainActivity.kt
@@ -1,9 +1,8 @@
-package com.sadegh.calculator
+package com.sadegh.calculator.presentation
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import com.sadegh.calculator.homeScreen.HomeScreen
 import com.sadegh.calculator.ui.theme.CalculatorTheme
 
 class MainActivity : ComponentActivity() {
